# Push limit into aggregation for DISTINCT ... LIMIT queries