… datafusion/core/src/physical_optimizer/optimizer.rs
… datafusion/core/src/physical_optimizer/topk_aggregation.rs
… datafusion/core/tests/sql/group_by.rs
… datafusion/physical-plan/src/aggregates/mod.rs
… datafusion/sqllogictest/test_files/aggregate.slt
… datafusion/sqllogictest/test_files/explain.slt
… datafusion/sqllogictest/test_files/information_schema.slt
… docs/source/user-guide/configs.md