@msirek 2023-11-08
> I have a suggestion on how to simplify this code in https://github.com/msirek/arrow-datafusion/pull/1

Thanks! I've included those commits in this PR.

> BTW I would love to know anything you can share about your usecase and what you are doing with DataFusion

I'm not really a DataFusion user. I just have a personal interest in query optimization and DataFusion looks pretty neat.