@alamb 2023-11-08
<img width="747" alt="Screenshot 2023-11-08 at 5 21 38 PM" src="https://github.com/apache/arrow-datafusion/assets/490673/3e57180f-656e-4739-ad8e-0279cbee004b">
🤔  this branch has some conflicts that need to be fixed

Perhaps #8004